do DELETE/UPDATEs only on install, not on upgrade

+++ src/skins/mkr_postinst_monster.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+PrevVer="$2"
+
 . /usr/pluto/bin/Config_Ops.sh
 . /usr/pluto/bin/SQL_Ops.sh
 
@@ -16,15 +18,17 @@ cp /usr/pluto/orbiter/skins/Monster/bootsplash/*.cfg /etc/bootsplash/themes/plut
 cp /usr/pluto/orbiter/skins/Monster/bootsplash/*.jpg /etc/bootsplash/themes/pluto/images
 /var/lib/dpkg/info/bootsplash.postinst configure
 
-# Get rid of the default room / user since we'll do this in the wizard
-Q="DELETE FROM Room WHERE PK_Room=1"
-RunSQL "$Q"
+if [[ -z "$PrevVer" ]]; then
+	# Get rid of the default room / user since we'll do this in the wizard
+	Q="DELETE FROM Room WHERE PK_Room=1"
+	RunSQL "$Q"
 
-Q="UPDATE Device SET FK_Room=NULL WHERE FK_Room=1"
-RunSQL "$Q"
+	Q="UPDATE Device SET FK_Room=NULL WHERE FK_Room=1"
+	RunSQL "$Q"
 
-Q="DELETE FROM Users WHERE PK_Users=1"
-RunSQL "$Q"
+	Q="DELETE FROM Users WHERE PK_Users=1"
+	RunSQL "$Q"
 
-Q="DELETE FROM Installation_Users WHERE FK_Users=1"
-RunSQL "$Q"
+	Q="DELETE FROM Installation_Users WHERE FK_Users=1"
+	RunSQL "$Q"
+fi
